refactor(pdf-viewer): extract drag-to-pan logic into useOverflowDrag composable
Move overflow drag and resize observer logic out of FzPdfViewer into
src/composables/useOverflowDrag.ts. Composable accepts a container ref
and returns a cursorClass computed for the template.
